requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LangChain & LangGraph for agent workflow
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.1.0
langchain-core>=0.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Utilities
python-dateutil>=2.8.2
httpx>=0.27.0  # For web search via Tavily API and Deep Research integration
tenacity>=8.2.0  # Retry logic with exponential backoff
